e-threader/requirements.txt
discord.py
EbookLib
pony
asyncpraw
Markdown
jinja2
